requirements.txt
aiosmtplib==2.0.2
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgi-correlation-id==4.3.1
asyncpg==0.29.0
bcrypt==4.1.3
billiard==4.2.0
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
celery==5.4.0
certifi==2024.6.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
loguru==0.7.2
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
passlib==1.7.4
prompt-toolkit==3.0.47
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.3.3
pygments==2.18.0
pyjwt==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
redis==5.0.6
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn[standard]==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.3
